feat(ever-admin): improved Payments step, export duplicate code and initialize PayPal configuration form EV-1192
