+++ admin/website-angular/src/app/pages/+setup/+merchants/components/payments/payment-gateways/payment-gateways.component.html
@@ -1,6 +1,11 @@
 <ea-stripe-gateway
 	#stripeGateway
+	[currenciesCodes]="currenciesCodes"
 	[companyBrandLogo]="warehouseLogo"
 	[warehouseCountry]="warehouseCountry"
 ></ea-stripe-gateway>
-<ea-payPal-gateway #payPalGateway></ea-payPal-gateway>
+<ea-payPal-gateway
+	#payPalGateway
+	[currenciesCodes]="currenciesCodes"
+	[warehouseCountry]="warehouseCountry"
+></ea-payPal-gateway>
